Admin update item

+++ resources/lablist_and_checkout.js
@@ -140,12 +140,14 @@ async function populateItemDetails(labLocation, itemId) {
   };
 
   const deleteButton = document.getElementById("delete-button");
-  deleteButton.onclick = function() {
-    const labName = document.getElementById('lab-name').textContent;
-    const currentPage = parseInt(document.querySelector('#pagination button[disabled]') === null ? 1 : document.querySelector('#pagination button[disabled]').textContent);
-    const searchValue = document.getElementById('search').value;
-    delete_item(labName, itemId, currentPage, searchValue);
-  };
+  if(deleteButton!==null){
+    deleteButton.onclick = function() {
+      const labName = document.getElementById('lab-name').textContent;
+      const currentPage = parseInt(document.querySelector('#pagination button[disabled]') === null ? 1 : document.querySelector('#pagination button[disabled]').textContent);
+      const searchValue = document.getElementById('search').value;
+      delete_item(labName, itemId, currentPage, searchValue);
+    };
+  }
 }
 
 function populate() {
